LPS-118271 Refactor drag and drop
