+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/utils/dragAndDrop/checkAllowedChild.js
@@ -1,0 +1,84 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+/**
+ * Checks if the given child can be nested inside given parent
+ * @param {object} child
+ * @param {object} parent
+ * @return {boolean}
+ */
+
+import {LAYOUT_DATA_ITEM_TYPES} from '../../config/constants/layoutDataItemTypes';
+
+const LAYOUT_DATA_ALLOWED_CHILDREN_TYPES = {
+	[LAYOUT_DATA_ITEM_TYPES.root]: [
+		LAYOUT_DATA_ITEM_TYPES.collection,
+		LAYOUT_DATA_ITEM_TYPES.dropZone,
+		LAYOUT_DATA_ITEM_TYPES.container,
+		LAYOUT_DATA_ITEM_TYPES.row,
+		LAYOUT_DATA_ITEM_TYPES.fragment,
+	],
+	[LAYOUT_DATA_ITEM_TYPES.collection]: [],
+	[LAYOUT_DATA_ITEM_TYPES.collectionItem]: [
+		LAYOUT_DATA_ITEM_TYPES.container,
+		LAYOUT_DATA_ITEM_TYPES.row,
+		LAYOUT_DATA_ITEM_TYPES.fragment,
+	],
+	[LAYOUT_DATA_ITEM_TYPES.dropZone]: [],
+	[LAYOUT_DATA_ITEM_TYPES.container]: [
+		LAYOUT_DATA_ITEM_TYPES.collection,
+		LAYOUT_DATA_ITEM_TYPES.container,
+		LAYOUT_DATA_ITEM_TYPES.dropZone,
+		LAYOUT_DATA_ITEM_TYPES.row,
+		LAYOUT_DATA_ITEM_TYPES.fragment,
+	],
+	[LAYOUT_DATA_ITEM_TYPES.row]: [LAYOUT_DATA_ITEM_TYPES.column],
+	[LAYOUT_DATA_ITEM_TYPES.column]: [
+		LAYOUT_DATA_ITEM_TYPES.collection,
+		LAYOUT_DATA_ITEM_TYPES.container,
+		LAYOUT_DATA_ITEM_TYPES.dropZone,
+		LAYOUT_DATA_ITEM_TYPES.row,
+		LAYOUT_DATA_ITEM_TYPES.fragment,
+	],
+	[LAYOUT_DATA_ITEM_TYPES.fragment]: [],
+	[LAYOUT_DATA_ITEM_TYPES.fragmentDropZone]: [
+		LAYOUT_DATA_ITEM_TYPES.collection,
+		LAYOUT_DATA_ITEM_TYPES.dropZone,
+		LAYOUT_DATA_ITEM_TYPES.container,
+		LAYOUT_DATA_ITEM_TYPES.row,
+		LAYOUT_DATA_ITEM_TYPES.fragment,
+	],
+};
+
+export default function checkAllowedChild(child, parent, layoutDataRef) {
+	const parentIsInsideCollection = (function checkItemInsideCollection(item) {
+		if (item.type === LAYOUT_DATA_ITEM_TYPES.collection) {
+			return true;
+		}
+		else if (item.parentId) {
+			return checkItemInsideCollection(
+				layoutDataRef.current.items[item.parentId]
+			);
+		}
+	})(parent);
+
+	if (
+		parentIsInsideCollection &&
+		child.type === LAYOUT_DATA_ITEM_TYPES.collection
+	) {
+		return false;
+	}
+
+	return LAYOUT_DATA_ALLOWED_CHILDREN_TYPES[parent.type].includes(child.type);
+}
