+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/utils/dragAndDrop/itemIsAncestor.js
@@ -1,0 +1,35 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+/**
+ * Checks if the given parent is indeed parent of given child.
+ * @param {object} child
+ * @param {object} parent
+ * @param {object} layoutDataRef
+ * @return {boolean}
+ */
+
+export default function itemIsAncestor(child, parent, layoutDataRef) {
+	if (child && parent) {
+		return child.itemId !== parent.itemId
+			? itemIsAncestor(
+					child,
+					layoutDataRef.current.items[parent.parentId],
+					layoutDataRef
+			  )
+			: true;
+	}
+
+	return false;
+}
