+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/fragments-widgets/components/TabItem.js
@@ -25,7 +25,7 @@ import {useDispatch, useSelector} from '../../../app/store/index';
 import addFragment from '../../../app/thunks/addFragment';
 import addItem from '../../../app/thunks/addItem';
 import addWidget from '../../../app/thunks/addWidget';
-import {useDragSymbol} from '../../../app/utils/useDragAndDrop';
+import {useDragSymbol} from '../../../app/utils/dragAndDrop/useDragAndDrop';
 
 export default function TabItem({item}) {
 	const dispatch = useDispatch();
